pyproject.toml
[tool.poetry]
name = "pythonproject10"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.4"
setuptools = "^75.7.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.10.0"
mypy = "^1.14.1"
requests = "^2.32.3"
pandas = "^2.2.3"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
